+++ src/ResourceGraph/ResourceGraph.Autorest/exports/ProxyCmdletDefinitions.ps1
@@ -53,7 +53,7 @@ INPUTOBJECT <IResourceGraphIdentity>: Identity Parameter
   [ResourceName <String>]: The name of the Graph Query resource.
   [SubscriptionId <String>]: The Azure subscription Id.
 .Link
-https://docs.microsoft.com/en-us/powershell/module/az.resourcegraph/get-azresourcegraphquery
+https://learn.microsoft.com/en-us/powershell/module/az.resourcegraph/get-azresourcegraphquery
 #>
 function Get-AzResourceGraphQuery {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.ResourceGraph.Models.Api20180901Preview.IGraphQueryResource])]
@@ -218,7 +218,7 @@ INPUTOBJECT <IResourceGraphIdentity>: Identity Parameter
   [ResourceName <String>]: The name of the Graph Query resource.
   [SubscriptionId <String>]: The Azure subscription Id.
 .Link
-https://docs.microsoft.com/en-us/powershell/module/az.resourcegraph/remove-azresourcegraphquery
+https://learn.microsoft.com/en-us/powershell/module/az.resourcegraph/remove-azresourcegraphquery
 #>
 function Remove-AzResourceGraphQuery {
 [OutputType([System.Boolean])]
@@ -384,7 +384,7 @@ Microsoft.Azure.PowerShell.Cmdlets.ResourceGraph.Models.IResourceGraphIdentity
 .Outputs
 Microsoft.Azure.PowerShell.Cmdlets.ResourceGraph.Models.Api20180901Preview.IGraphQueryResource
 .Link
-https://docs.microsoft.com/en-us/powershell/module/az.resourcegraph/new-azresourcegraphquery
+https://learn.microsoft.com/en-us/powershell/module/az.resourcegraph/new-azresourcegraphquery
 #>
 function New-AzResourceGraphQuery {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.ResourceGraph.Models.Api20180901Preview.IGraphQueryResource])]
@@ -575,7 +575,7 @@ INPUTOBJECT <IResourceGraphIdentity>: Identity Parameter
   [ResourceName <String>]: The name of the Graph Query resource.
   [SubscriptionId <String>]: The Azure subscription Id.
 .Link
-https://docs.microsoft.com/en-us/powershell/module/az.resourcegraph/update-azresourcegraphquery
+https://learn.microsoft.com/en-us/powershell/module/az.resourcegraph/update-azresourcegraphquery
 #>
 function Update-AzResourceGraphQuery {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.ResourceGraph.Models.Api20180901Preview.IGraphQueryResource])]