+++ src/ContainerInstance/Changelog.md
@@ -20,6 +20,10 @@
 
 ## Upcoming Release
 
+## Version 3.2.0
+* Added `priority` property to Container Group properties
+* Added `Confidential` sku type to Container Group Skus
+
 ## Version 3.1.0
 * Supported empty directory volume and secret volume for creating container group [#17410]