+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,10 +18,13 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Supported Blob Tier Cold
+* Supported new Blob Tier Cold
     - `Copy-AzStorageBlob`
     - `Set-AzStorageBlobContent` 
     - `Start-AzStorageBlobCopy`
+
+## Version 5.4.1
+* Updated Azure.Core to 1.28.0.
 
 ## Version 5.4.0
 * Added a warning message for the upcoming breaking change when creating a Storage account